Conflicts:
	h2o-core/src/main/java/water/api/RapidsHandler.java
	h2o-core/src/main/java/water/api/RapidsV99.java
	h2o-core/src/main/java/water/api/RequestServer.java
	h2o-core/src/main/java/water/fvec/Frame.java
	h2o-py/h2o/__init__.py
	h2o-py/h2o/frame.py
	h2o-py/h2o/h2o.py
	h2o-py/h2o/h2o_model_builder.py
	h2o-py/h2o/model/__init__.py
	h2o-py/tests/pyunit_user.py
	h2o-r/H2O_Load.R
	h2o-r/h2o-package/R/ast.R
	h2o-r/h2o-package/R/astfun.R
	h2o-r/h2o-package/R/communication.R
	h2o-r/h2o-package/R/connection.R
	h2o-r/h2o-package/R/deeplearning.R
	h2o-r/h2o-package/R/exec.R
	h2o-r/h2o-package/R/export.R
	h2o-r/h2o-package/R/frame.R
	h2o-r/h2o-package/R/glm.R
	h2o-r/h2o-package/R/import.R
	h2o-r/h2o-package/R/pca.R

+++ h2o-algos/src/main/java/hex/naivebayes/NaiveBayes.java
@@ -3,6 +3,7 @@ package hex.naivebayes;
 import hex.*;
 import hex.schemas.ModelBuilderSchema;
 import hex.schemas.NaiveBayesV3;
+import jsr166y.CountedCompleter;
 import water.*;
 import water.exceptions.H2OModelBuilderIllegalArgumentException;
 import water.fvec.Chunk;
@@ -16,6 +17,7 @@ import water.util.TwoDimTable;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * Naive Bayes
@@ -26,15 +28,17 @@ import java.util.List;
  * @author anqi_fu
  *
  */
-public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBayesModel.NaiveBayesParameters,NaiveBayesModel.NaiveBayesOutput> {
+public class NaiveBayes extends ModelBuilder<NaiveBayesModel,NaiveBayesModel.NaiveBayesParameters,NaiveBayesModel.NaiveBayesOutput> {
   @Override
   public ModelBuilderSchema schema() {
     return new NaiveBayesV3();
   }
 
+  public boolean isSupervised(){return true;}
+
   @Override
   public Job<NaiveBayesModel> trainModel() {
-    return start(new NaiveBayesDriver(), 0);
+    return start(new NaiveBayesDriver(), 6);
   }
 
   @Override
@@ -92,11 +96,13 @@ public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBaye
 
   class NaiveBayesDriver extends H2O.H2OCountedCompleter<NaiveBayesDriver> {
 
-    public void computeStatsFillModel(NaiveBayesModel model, DataInfo dinfo, NBTask tsk) {
+    public boolean computeStatsFillModel(NaiveBayesModel model, DataInfo dinfo, NBTask tsk) {
       model._output._levels = _response.domain();
       model._output._rescnt = tsk._rescnt;
       model._output._ncats = dinfo._cats;
 
+      if(!isRunning(_key)) return false;
+      update(1, "Initializing arrays for model statistics");
       // String[][] domains = dinfo._adaptedFrame.domains();
       String[][] domains = model._output._domains;
       double[] apriori = new double[tsk._nrescat];
@@ -106,6 +112,8 @@ public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBaye
         pcond[i] = new double[tsk._nrescat][ncnt];
       }
 
+      if(!isRunning(_key)) return false;
+      update(1, "Computing probabilities for categorical cols");
       // A-priori probability of response y
       for(int i = 0; i < apriori.length; i++)
         apriori[i] = ((double)tsk._rescnt[i] + _parms._laplace)/(tsk._nobs + tsk._nrescat * _parms._laplace);
@@ -120,6 +128,8 @@ public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBaye
         }
       }
 
+      if(!isRunning(_key)) return false;
+      update(1, "Computing mean and standard deviation for numeric cols");
       // Mean and standard deviation of numeric predictor x_j for every level of response y
       for(int col = 0; col < dinfo._nums; col++) {
         for(int i = 0; i < pcond[0].length; i++) {
@@ -163,8 +173,10 @@ public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBaye
       Arrays.fill(colFormats, "%5f");
       model._output._apriori = new TwoDimTable("A Priori Response Probabilities", null, new String[1], _response.domain(), colTypes, colFormats, "",
               new String[1][], new double[][] {apriori});
-
       model._output._model_summary = createModelSummaryTable(model._output);
+
+      if(!isRunning(_key)) return false;
+      update(1, "Scoring and computing metrics on training data");
       if (_parms._compute_metrics) {
         model.score(_parms.train()).delete(); // This scores on the training data and appends a ModelMetrics
         ModelMetricsSupervised mm = DKV.getGet(model._output._model_metrics[model._output._model_metrics.length - 1]);
@@ -172,32 +184,37 @@ public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBaye
       }
 
       // At the end: validation scoring (no need to gather scoring history)
+      if(!isRunning(_key)) return false;
+      update(1, "Scoring and computing metrics on validation data");
       if (_valid != null) {
         Frame pred = model.score(_parms.valid()); //this appends a ModelMetrics on the validation set
         model._output._validation_metrics = DKV.getGet(model._output._model_metrics[model._output._model_metrics.length - 1]);
         pred.delete();
       }
+
+      return true;
     }
 
-    @Override
-    protected void compute2() {
+    @Override protected void compute2() {
       NaiveBayesModel model = null;
       DataInfo dinfo = null;
 
       try {
-        init(true);
-        if (error_count() > 0) throw H2OModelBuilderIllegalArgumentException.makeFromBuilder(NaiveBayes.this);
+        init(true);   // Initialize parameters
         _parms.read_lock_frames(NaiveBayes.this); // Fetch & read-lock input frames
-        dinfo = new DataInfo(Key.make(), _train, _valid, 1, false, DataInfo.TransformType.NONE, DataInfo.TransformType.NONE, true, false);
+        if (error_count() > 0) throw H2OModelBuilderIllegalArgumentException.makeFromBuilder(NaiveBayes.this);
+        dinfo = new DataInfo(Key.make(), _train, _valid, 1, false, DataInfo.TransformType.NONE, DataInfo.TransformType.NONE, true, false, false, false);
 
         // The model to be built
         model = new NaiveBayesModel(dest(), _parms, new NaiveBayesModel.NaiveBayesOutput(NaiveBayes.this));
         model.delete_and_lock(_key);
         _train.read_lock(_key);
 
-        NBTask tsk = new NBTask(dinfo, _response.cardinality()).doAll(dinfo._adaptedFrame);
-        computeStatsFillModel(model, dinfo, tsk);
-        model.update(_key);
+        update(1, "Begin distributed Naive Bayes calculation");
+        NBTask tsk = new NBTask(_key, dinfo, _response.cardinality()).doAll(dinfo._adaptedFrame);
+        if (computeStatsFillModel(model, dinfo, tsk))
+          model.update(_key);
+
         done();
       } catch (Throwable t) {
         Job thisJob = DKV.getGet(_key);
@@ -254,6 +271,7 @@ public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBaye
   //             If response y = NA, skip counting row entirely in all calculations
   // H2O's method: Just skip all rows where any x_j = NA or y = NA. Should be more memory-efficient, but results incomparable with R.
   private static class NBTask extends MRTask<NBTask> {
+    final protected Key _jobKey;
     final DataInfo _dinfo;
     final String[][] _domains;  // Domains of the training frame
     final int _nrescat;         // Number of levels for the response y
@@ -264,7 +282,8 @@ public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBaye
     public int[/*npreds*/][/*nrescat*/][] _jntcnt;  // For each categorical predictor, joint count of response and predictor levels
     public double[/*npreds*/][/*nrescat*/][] _jntsum; // For each numeric predictor, sum and squared sum of entries for every response level
 
-    public NBTask(DataInfo dinfo, int nres) {
+    public NBTask(Key jobKey, DataInfo dinfo, int nres) {
+      _jobKey = jobKey;
       _dinfo = dinfo;
       _nrescat = nres;
       _domains = dinfo._adaptedFrame.domains();
@@ -274,6 +293,9 @@ public class NaiveBayes extends SupervisedModelBuilder<NaiveBayesModel,NaiveBaye
     }
 
     @Override public void map(Chunk[] chks) {
+      if(_jobKey != null && !isRunning(_jobKey)) {
+        throw new JobCancelledException();
+      }
       _nobs = 0;
       _rescnt = new int[_nrescat];
